chore(frontend): fix SonarQube token usage, Node heap size, and baseline-browser-mapping
- Switch Jenkins SonarQube to sonar.token
- Add -Dsonar.javascript.node.maxspace=2048
- Add baseline-browser-mapping@2.9.19 as devDependency
- Use npm ci with fallback to npm install
- Sync package-lock.json
