--- expensive-tracker-frontend/src/app/Navigation/components/settings/contact/contact.component.html
@@ -1,1 +1,0 @@
-<p>contact works!</p>
