--- expensive-tracker-frontend/src/app/Navigation/components/settings/contact/contact.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-contact',
-  imports: [],
-  templateUrl: './contact.component.html',
-  styleUrl: './contact.component.css'
-})
-export class ContactComponent {
-
-}
