--- expensive-tracker-frontend/src/app/services/toastmsg.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ToastmsgService } from './toastmsg.service';
-
-describe('ToastmsgService', () => {
-  let service: ToastmsgService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ToastmsgService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
