--- expensive-tracker-frontend/src/app/shared/components/toast/toast.component.css
@@ -1,13 +1,0 @@
-@keyframes slideIn {
-  from {
-    transform: translateX(100%);
-    opacity: 0;
-  }
-  to {
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
-.animate-slideIn {
-  animation: slideIn 0.3s ease-out;
-}
