--- expensive-tracker-frontend/src/app/shared/components/toast/toast.component.html
@@ -1,13 +1,0 @@
-<div class="fixed bottom-4 right-4 z-50 space-y-2">
-  <div *ngFor="let toast of toasts"
-       class="flex items-center p-4 rounded-lg shadow-lg animate-slide-in-up"
-       [class]="getToastClass(toast.type)">
-    <fa-icon [icon]="getIcon(toast.type)" class="flex-shrink-0 w-5 h-5 mr-3"></fa-icon>
-    <p class="text-sm font-medium">{{ toast.message }}</p>
-    <button type="button"
-            (click)="toastService.remove(toast.id)"
-            class="ml-auto -mx-1.5 -my-1.5 rounded-lg p-1.5 inline-flex h-8 w-8 hover:bg-white/20 focus:ring-2 focus:ring-white/50">
-      <fa-icon [icon]="faTimes"></fa-icon>
-    </button>
-  </div>
-</div>
